requirements.txt
streamlit==1.34.0
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
google-cloud-bigquery==3.21.0
python-dotenv==1.0.0
db-dtypes==1.1.1
gspread==5.12.4
google-auth==2.23.4